producers/requirements.txt
attrs==23.1.0
avro==1.12.0
blessed==1.20.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
confluent-kafka==2.5.3
contourpy==1.0.7
cycler==0.11.0
editor==1.6.6
fastavro==1.9.7
Fiona==1.9.4.post1
fonttools==4.39.4
geographiclib==2.0
geopandas==0.13.0
geopy==2.3.0
idna==3.4
inquirer==3.4.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.4
matplotlib==3.7.1
networkx==3.1
numpy==1.24.3
osmnx==1.3.1.post0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
prettytable==3.11.0
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
readchar==4.2.0
referencing==0.35.1
requests==2.31.0
rpds-py==0.20.0
runs==1.2.2
shapely==2.0.1
six==1.16.0
tzdata==2023.3
urllib3==2.0.2
wcwidth==0.2.13
xmod==1.8.1
